Removed SignIn.Java and moved it's methods to SignInActivity.Java Also implemented adding players to database.
